ml-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision & ML
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.196   # YOLOv8 (uniform detection)

# Face recognition (modern, no dlib)
deepface==0.0.93

# Utilities
python-dateutil==2.8.2
